Adiciona teste para a pesquisa de carrinho via QueryParam
